backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.6
groq>=0.9.0
python-dotenv>=1.0.0
networkx>=3.3
